Added semaphore timeouts.

+++ include/sc.hpp
@@ -4,7 +4,8 @@
  * Copyright (C) 2009-2011 Udo Steinberg <udo@hypervisor.org>
  * Economic rights: Technische Universitaet Dresden (Germany)
  *
- * Copyright (C) 2012 Udo Steinberg, Intel Corporation.
+ * Copyright (C) 2012-2013 Udo Steinberg, Intel Corporation.
+ * Copyright (C) 2014 Udo Steinberg, FireEye, Inc.
  *
  * This file is part of the NOVA microhypervisor.
  *
@@ -32,11 +33,11 @@ class Sc : public Kobject
         Refptr<Ec> const ec;
         unsigned const cpu;
         unsigned const prio;
-        unsigned const budget;
+        uint64 const budget;
         uint64 time;
 
     private:
-        unsigned left;
+        uint64 left;
         Sc *prev, *next;
         uint64 tsc;
